Add files via upload
In order for the main page to work you need to upload the 2 html in the spiffs through the website. First upload the main code. Then open dipae.local/upload and chose the 2 html file to upload on spiffs. When you finish hit the home button to redirect you to the home page.